Add a function to find running conductor workflows
