+++ packages/e2e-test/helpers/ConductorHelper.ts
@@ -1,5 +1,21 @@
-export default {
-  CONDUCTOR_URL: process.env.CONDUCTOR_URL ?? "http://localhost:5002/api",
-  CONDUCTOR_USER: process.env.CONDUCTOR_USER ?? "bichard",
-  CONDUCTOR_PASSWORD: process.env.CONDUCTOR_PASSWORD ?? "password"
+import createConductorClient from "@moj-bichard7/common/conductor/createConductorClient"
+
+const client = createConductorClient()
+
+export const findRunningConductorWorkflowIds = async (): Promise<string[]> => {
+  const searchResult = await client.workflowResource.search1(
+    undefined,
+    0,
+    100,
+    "startTime:DESC",
+    undefined,
+    "status='RUNNING'"
+  )
+
+  return searchResult.results?.map((workflow) => workflow.workflowId!) || []
+}
+
+export const terminateConductorWorkflows = async () => {
+  const idsToTerminate = await findRunningConductorWorkflowIds()
+  await Promise.all(idsToTerminate.map((id) => client.workflowResource.terminate1(id, "Termination by test script")))
 }
